Code refactor with useWindowFocusPollingEffect.
